event sorting is now easier to read and properly typed
